+++ js/editor.js
@@ -35,9 +35,12 @@
         OCA.Onlyoffice.template = $("#iframeEditor").data("template");
         OCA.Onlyoffice.inframe = !!$("#iframeEditor").data("inframe");
         OCA.Onlyoffice.anchor = $("#iframeEditor").attr("data-anchor");
+        OCA.Onlyoffice.currentWindow = window;
 
         if (OCA.Onlyoffice.inframe) {
             OCA.Onlyoffice.faviconBase = $('link[rel="icon"]').attr("href");
+            OCA.Onlyoffice.currentWindow = window.parent;
+            OCA.Onlyoffice.titleBase = OCA.Onlyoffice.currentWindow.document.title;
         }
 
         if (!OCA.Onlyoffice.fileId && !OCA.Onlyoffice.shareToken) {
@@ -111,7 +114,7 @@
 
                         if (docIsChanged !== event.data) {
                             var titleChange = function () {
-                                window.document.title = config.document.title + (event.data ? " *" : "") + " - " + oc_defaults.title;
+                                OCA.Onlyoffice.currentWindow.document.title = config.document.title + (event.data ? " *" : "") + " - " + oc_defaults.title;
                                 docIsChanged = event.data;
                             };
 
@@ -144,9 +147,13 @@
                         config.events.onRequestInsertImage = OCA.Onlyoffice.onRequestInsertImage;
                         config.events.onRequestMailMergeRecipients = OCA.Onlyoffice.onRequestMailMergeRecipients;
                         config.events.onRequestCompareFile = OCA.Onlyoffice.onRequestCompareFile;
-                        config.events.onRequestUsers = OCA.Onlyoffice.onRequestUsers;
                         config.events.onRequestSendNotify = OCA.Onlyoffice.onRequestSendNotify;
+                        config.events.onRequestReferenceData = OCA.Onlyoffice.onRequestReferenceData;
                         config.events.onMetaChange = OCA.Onlyoffice.onMetaChange;
+
+                        if (OC.currentUser) {
+                            config.events.onRequestUsers = OCA.Onlyoffice.onRequestUsers;
+                        }
 
                         if (!OCA.Onlyoffice.filePath) {
                             OCA.Onlyoffice.filePath = config._file_path;
@@ -205,7 +212,7 @@
     };
 
     OCA.Onlyoffice.onRequestHistoryData = function (event) {
-        var version = event.data; 
+        var version = event.data;
 
         $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/version?fileId={fileId}&version={version}",
             {
@@ -476,7 +483,7 @@
     };
 
     OCA.Onlyoffice.onRequestUsers = function (event) {
-        $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/users?fileId={fileId}", 
+        $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/users?fileId={fileId}",
         {
             fileId: OCA.Onlyoffice.fileId || 0
         }),
@@ -485,6 +492,25 @@
                 "users": response
             });
         });
+    };
+
+    OCA.Onlyoffice.onRequestReferenceData = function (event) {
+        let referenceData = event.data.referenceData;
+        let path = event.data.path;
+
+        $.post(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/reference"),
+            {
+                referenceData: referenceData,
+                path: path
+            },
+            function onSuccess(response) {
+                if (response.error) {
+                    OCA.Onlyoffice.showMessage(response.error, "error");
+                    return;
+                }
+
+                OCA.Onlyoffice.docEditor.setReferenceData(response);
+            });
     };
 
     OCA.Onlyoffice.onRequestSendNotify = function (event) {